New design using scss
