+++ src/components/Card/Card.scss
@@ -1,0 +1,158 @@
+@use '../../styles//defaultColors.scss' as colors;
+@use '../../styles/defaultFonts' as fonts;
+@use '../../styles/variables' as variables;
+
+.mainCardContainer {
+    display: flex;
+    align-items: center;
+    box-shadow: 0 3px 6px colors.$border-shadow;
+    background-color: colors.$white;
+    overflow: hidden;
+    border-radius: 3px;
+
+    &--small {
+        align-items: center;
+        height: 5rem;
+    }
+
+    &--large {
+        flex-direction: column;
+        width: 14.0rem;
+    }
+
+    .cardContainer {
+        transition: all 0.1s ease-in-out;
+        display: flex;
+        padding: 0.75rem;
+        overflow: hidden;
+
+        &--small {
+            align-items: center;
+            flex-grow: 1;
+            gap: 1rem;
+        }
+
+        &--large {
+            flex-direction: column;
+            width: 12.5rem;
+            gap: 0.625rem;
+        }
+
+        &:hover {
+            background-color: colors.$light-blue-transparent;
+        }
+
+        &--active {
+            background-color: colors.$bright-blue;
+            color: white;
+        }
+
+        
+
+
+        > .cardImage {
+            object-fit: cover;
+
+            &--large {
+                height: 7rem;
+                width: 12.5rem;
+                align-self: center;
+            }
+
+            &--small {
+                min-width: 3.5rem;
+                min-height: 3.5rem;
+                max-width: 3.5rem;
+                max-height: 3.5rem;
+            }
+        }
+
+        > .cardImagePlaceholder {
+            fill: colors.$gray-incomplete;
+
+            > rect {
+                width: inherit;
+                height: inherit;
+            }
+
+            &--large {
+                height: 7rem;
+                width: 12.5rem;
+                align-self: center;
+            }
+
+            &--small {
+                width: 3.5rem;
+                height: 3.5rem;
+            }
+        }
+
+        > .cardWrapper {
+            display: flex;
+            flex-direction: column;
+            gap: 0.5rem;
+            font-family: fonts.$font-family;
+            font-weight: fonts.$regular;
+            width: 100%;
+
+            &--large {
+                font-size: fonts.$px18;
+            }
+
+            &--small {
+                font-size: fonts.$px14;
+            }
+
+            > .cardTitle {
+                overflow: hidden;
+                white-space: nowrap;
+                text-overflow: ellipsis;
+
+                &--small {
+                    max-width: 50%;
+                }
+            }
+
+            > .cardCaption {
+                font-family: fonts.$font-family;
+                font-size: fonts.$px11;
+                font-weight: fonts.$regular;
+                color: colors.$gray-text;
+                margin: 0;
+            }
+
+            > .tagWrapper {
+                display: flex;
+                flex-wrap: wrap;
+                gap: 5px;
+            }
+        }
+
+
+
+    }
+
+    > .cardChildrenWrapper {
+        transition: all 0.1s ease-in-out;
+        height: 5rem;
+        aspect-ratio: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background-color: white;
+        color: colors.$dark-bright-blue;
+        cursor: pointer;
+
+        > svg {
+            transition: all 0.1s ease-in-out;
+        }
+
+        &:hover {
+            background-color: colors.$light-blue-transparent;
+            > svg {
+                transform: scale(1.3);
+            }
+        }
+    }
+
+}
