+++ src/stories/FilterGroup.stories.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import React, { useState } from 'react';
 
 import { FilterGroup } from '../components';
